# Parallelize requests for better performance