J'ai threadé les requests pour améliorer les performances. C'est pas encore top je pense qu'on peut mieux faire il faut que je rajoute du log pour voir où le temps passe. En moyenne je passe de 9s à 7s.
Je m'attendrais plutot à 1s ou 2s.