@Batalex 2017-10-17
Merci de ta contribution ! Je te laisse le temps d'explorer tout ça avant de merge :)
Petite question, est-ce que tu as testé de rendre simplement asynchrone l'obtention des données avec du `async`, `await` ?