@bigsool 2017-10-17
Python c'est pas forcément le langage que je maitrise le mieux donc à prendre avec des pincettes ce que je vais dire.

`async` et `await` ne vont pas rendre un traitement qui est synchrone , asynchrone.
Je me suis basé sur http://docs.python-requests.org/en/master/user/advanced/#blocking-or-non-blocking qui indique `requests-threads` ou `requests-futures`  (`grequests` est déprécié). J'ai choisi `requests-futures` parce que j'étais plus à l'aise, mais on pourrait utiliser `requests-threads` avec du `async` `await`.
Une fois les requests remplacés, je me suis rendu compte que c'était plus simple et plus lisible en lançant des threads.

Donc si tu as une préférence pour `async` `await` je m'adapte c'est toi qui décides. Ce soir après le boulot je regarde si j'ai le temps. Je vais rajouter du log en mode debug je pense.