@ferjul17 2017-10-17
J'ai changé de machine et les performances sont bien meilleures. Je passe de 4-5s avant ma PR à 2s avec ma PR. Je ne pense pas changer quoi que ce soit d'autre donc.