@Batalex 2017-10-18
Très bien je merge. Au pire on pourra toujours changer par la suite mais j'aime bien cette solution